% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsPlainVanilla.R
\name{bsPlainVanillaOption}
\alias{bsPlainVanillaOption}
\title{Compute price and greeks of plain vanilla European options}
\usage{
bsPlainVanillaOption(date, forward, strike, expiry, vol, type = "call",
  discount = 1, output = "price", model = "lognormal")
}
\arguments{
\item{date}{Value date of the option.}

\item{forward}{Forward level(s) of the underlying.}

\item{strike}{strike(s) of the option.}

\item{expiry}{Expiry date(s).}

\item{vol}{Volatility(s).}

\item{type}{Option type - "call" or "put".}

\item{discount}{Discount factor.}

\item{output}{Output type, "price" or any other standard greeks.}

\item{model}{Model type, only lognormal is implemented.}
}
\value{
Price or greek estimate. The function is vectorized.
}
\description{
Compute price and greeks of plain vanilla European options (
calls and puts) using lognormal Black-76 model.
}
\details{
Black's formula is arguably more useful for pricing index options
compared to traditional Black-Scholes formula. The Black-76 option pricing
model uses the price of futures directly, instead of accreting the spot
index at the cost of carry. This is useful where the index futures trade
at prices below spot plus cost of carry, providing a model consistent with
market. In many places this is now standard to use Black-76 for index
options valuation for margining and fair value determination purposes. For
greeks the return values are as follows: 1) delta: this returns the
standard black-scholes delta, the first derivative of option price with
respect to underlying. 2) gamma: this returns the change in delta for a
1 percentage point change in underlying (averaged for up and down move) 3)
vega: this returns the change in option price for a absolute 1 percentage
points increase in volatility 4) theta: this returns the decay in option
price over one day (assumeing 260 business days in a year).
}

