% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vols.R
\name{bumpVol}
\alias{bumpVol}
\title{bump the level of volatility for scenario analysis}
\usage{
bumpVol(vol, bump = 0.01)
}
\arguments{
\item{vol}{Object returned from earlier calibration. Either a quadvol or a
sabrvol object.}

\item{bump}{Bump amount. This is added to the ATM vol level}
}
\value{
A volatility object of class quadvol or sabrvol
}
\description{
Returns the volatility with a bumped up ATM vol levels
}
\seealso{
\code{\link{calibrate}} for calibration and models
}

