% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vols.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate vols (SABR or quadratic smile fit) from market prices}
\usage{
calibrate(options, valuedate, model = "quadratic", atmvol = 0,
  precision = 2, type = "delta")
}
\arguments{
\item{options}{dataframe with options chain details}

\item{valuedate}{valuation date for the calibration}

\item{model}{Volatility model, can be either sabr or quadratic}

\item{atmvol}{ATM volatility, if provided used in sabr calibration
(else ATM vol is estimated with a quadratic vol fit first)}

\item{precision}{degree of precision}

\item{type}{Type of fit in case of quadratic model (either delta or strike)}
}
\value{
Fitted volatility object. The object is either of class quadvol or
sabrvol depending upon the fit. The object includes atm vol level, and model
specific parameters. It also stores the valuation date of calibration. The
sabr model implements calibration of a lognormal SABR model (ref Hagan(2002)
"Managing smile risk" and West (2005) "Calibration of the SABR Model in
Illiquid Markets"). The model assumes a beta of 0 (lognormal behaviour) and
calibrates other parameters of SABR. The quadratic model is a simple two
parameters model of volatility smile, fitted with least square regression.
The calibration type delta fits this quadratic smile based on percentage
offset from ATM, while strike fits absolute deviation from ATM forward.
}
\description{
The first input is option chain data (dataframe) for a single
maturity, the expected columns include the forward, expiry (in date format),
strike, price and type (either call or put).
}

