% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distributionPricer}
\alias{distributionPricer}
\title{Price vanilla European options from a given underlying distribution}
\usage{
distributionPricer(forward, strike, voldist, type = "call", discount = 1,
  model = "lognormal")
}
\arguments{
\item{forward}{Forward level of the underlying.}

\item{strike}{Strike of the option.}

\item{voldist}{The distribution of the underlying (obtained using either
impliedDistribution or realizedDistribution function).}

\item{type}{The type of the option (call or put).}

\item{discount}{Discount factor.}

\item{model}{Model type, only lognormal is implemented.}
}
\value{
Price of the option.
}
\description{
Price vanilla European call or put options from a given
underlying distribution, using the first principle.
}
\details{
The underlying distribution is used to compute the payoff profile
of option and this is then integrated within the range to arrive at the
price of the option. All values are assumed to be zero outside the specified
distribution range.
#'@seealso \code{\link{impliedDistribution}}
}

