% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vols.R
\name{getVol}
\alias{getVol}
\title{get volatility at a given strike}
\usage{
getVol(vol, atm, strike)
}
\arguments{
\item{vol}{Object returned from earlier calibration. Either a quadvol or a
sabrvol object.}

\item{atm}{ATM forward(s)}

\item{strike}{Options strike(s)}
}
\value{
Black volatility at the given strike
}
\description{
Returns the volatility at a given strike. This function is
vectorized.
}
\seealso{
\code{\link{calibrate}} for calibration and models
}

