% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{impliedDistribution}
\alias{impliedDistribution}
\title{Extract implied distribution of the underlying from options prices}
\usage{
impliedDistribution(vol, n = 512, bounds = c(-0.2, 0.2),
  normalization = TRUE)
}
\arguments{
\item{vol}{A vol object (either sabrvol or quadvol) obtained by calibrating
to market prices of options.}

\item{n}{Number of division for the range of strikes to generate the
distribution.}

\item{bounds}{The range (in percentage) of the underlying to estimate the
distribution.}

\item{normalization}{A logical value. If true, the implied distribution is
normalized to sum up to 1, else not.}
}
\value{
A dataframe with strikes and corresponding probability density
function estimates.
}
\description{
Extract implied probability distribution from option prices
from a fitted vol model calibrated to market prices.
}
\details{
Extracting probability density from the implied volatility of
options is done using the standard method of taking the second derivatives
of option prices with respect to strikes. The second derivative is estimated
using the usual finite difference approach. The resulting distribution is
the option implied risk neutral distribution. The strikes of the underlying
of the distribution is always scaled so that the current forward is at 100.
}
\seealso{
\code{\link{calibrate}} for calibration and volatility models
}

